Handle LastPass entries that are categorized
For some reason the LaunchBar action was limited to only the `(none)` category of my LastPass entries. I'm not totally sure why it was failing, but this PR fixes that issue for me.

Also, when LastPass entries are categorized their `name` is a string with the categories prepended with `/`s (eg: for a `github.com` login catergorized under `Web Dev` you get: `Web Dev/github.com`) There can be sub categorizations too. In any case, on line 36 in this PR I used regex to get *just* the `name` for a cleaner UI, but this is optional.